[共通] 整理

+++ Siv3D/include/Siv3D/Vertex2D.hpp
@@ -17,7 +17,7 @@ namespace s3d
 {
 	struct Vertex2D
 	{
-		using IndexType = uint32;
+		using IndexType = uint16;
 
 		/// @brief 位置
 		Float2 pos;
